quiz update functionality and ui - done

+++ src/pages/api/quizzes/index.ts
@@ -8,13 +8,30 @@ export default async function handler(
   res: NextApiResponse
 ) {
   if (req.method === "DELETE") {
-    const { id } = req.query;
-    const quiz = await prisma.quizzes.delete({
-      where: {
-        id: Number(id),
-      },
-    });
-    return res.status(200).json(quiz);
+    const question = req.body;
+    if (question.quizId && question.questionId) {
+      const quiz = await prisma.quizzes.update({
+        where: {
+          id: Number(question.quizId),
+        },
+        data: {
+          questions: {
+            disconnect: {
+              id: question.questionId,
+            },
+          },
+        },
+      });
+      return res.status(200).json(quiz);
+    } else {
+      const { id } = req.query;
+      const quiz = await prisma.quizzes.delete({
+        where: {
+          id: Number(id),
+        },
+      });
+      return res.status(200).json(quiz);
+    }
   } else if (req.method === "GET") {
     const { id } = req.query;
     if (id) {
@@ -35,5 +52,17 @@ export default async function handler(
       });
       return res.status(200).json(quizzes);
     }
+  } else if (req.method === "PUT") {
+    const quizData = req.body;
+    console.log(quizData, "quizData");
+    const quiz = await prisma.quizzes.update({
+      where: {
+        id: Number(quizData.quizId),
+      },
+      data: {
+        name: quizData.quizName,
+      },
+    });
+    return res.status(200).json(quiz);
   }
 }
